Logica para componentes SecondEventView.js CalendarioEventoUser.js Elimine CrearEditarEvento.js
